pulp_tag: Make PulpUploader a utility class
Move PulpUploader out of post_push_to_pulp and into util.py so that it can
be shared with post_pulp_tag. Rename it into PulpHandler. In the process,
move the creation of the pulp logger out of PulpPushPlugin.__init() and
into PulpHandler.__init(). publish also stops being an argument to
PulpHandler and becomes an argument to push_tarball_to_pulp()

Signed-off-by: Mark Langsdorf <mlangsdo@redhat.com>