group_manifests: add koji_import support for manifest lists
If group_manifests created a manifest list, koji_import should create a dict
in 'extras': 'index' that contains the pullspec and tags for the manifest
list. Have koji_import check the return of group_manifests, and if it
got a list of manifest digests, have it set up the 'extras': 'index'
dictionary.

Signed-off-by: Mark Langsdorf <mlangsdo@redhat.com>