Pass components list in the workflow
Rather than using the plugin output of the 'rpmqa' plugin as the
place to store the components list, pass it by an explicit member
of the workflow. This allows other plugins to fill in the components
list rather than making it the exclusive domain of the 'rpmqa' plugin.

The output of the 'rpmqa' plugin is preserved so as not to hide the
component list from the logs.