group_manifests: add support for OCI manifests and images index
The code is rewritten to remove the dependency on manifest-tool, and
directly communicate with the registry via HTTP for manifest creation
in addition to, as before, tagging. The structure is changed to make
it clear that we can tag and group within a single registry (linking
blobs as necessary between repositories), but not between registries.

The test code gains a minimal mocked in-memory representation of a
Docker registry, and the group=true, and group=false cases are combined
since they now have much the same code paths, and need similar tests.`